refactor(ridesx): fold QDL platform flasher into ridesx package
Move QualcommFlasher and QDL manifest support from jumpstarter-driver-qualcomm
into jumpstarter_driver_ridesx.qdl. Extract shared TAC serial helpers without
changing RideSX fastboot/power behavior. Keep StreamingFlasherClient in core.
