--- docs/source/reference/package-apis/drivers/qualcomm.md
@@ -1,1 +1,0 @@
-../../../../../python/packages/jumpstarter-driver-qualcomm/README.md
