--- python/packages/jumpstarter-driver-qualcomm/jumpstarter_driver_qualcomm/__init__.py
@@ -1,1 +1,0 @@
-"""Jumpstarter driver for Qualcomm QDL firmware flashing and identification."""
