--- python/packages/jumpstarter-driver-qualcomm/pyproject.toml
@@ -1,48 +1,0 @@
-[project]
-name = "jumpstarter-driver-qualcomm"
-dynamic = ["version", "urls"]
-description = "Jumpstarter driver for Qualcomm QDL firmware flashing and identification"
-readme = "README.md"
-license = "Apache-2.0"
-authors = [{ name = "Jumpstarter Contributors" }]
-requires-python = ">=3.11"
-dependencies = [
-    "jumpstarter",
-    "jumpstarter-driver-composite",
-    "jumpstarter-driver-pyserial",
-    "pyyaml>=6.0.2",
-    "pydantic>=2.0",
-    "pexpect>=4.9",
-]
-
-[project.entry-points."jumpstarter.drivers"]
-QualcommFlasher = "jumpstarter_driver_qualcomm.driver:QualcommFlasher"
-
-[tool.hatch.version]
-source = "vcs"
-raw-options = { 'root' = '../../../' }
-
-[tool.hatch.metadata.hooks.vcs.urls]
-Homepage = "https://jumpstarter.dev"
-source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"
-
-[tool.pytest.ini_options]
-addopts = "--cov --cov-report=html --cov-report=xml"
-log_cli = true
-log_cli_level = "INFO"
-testpaths = ["jumpstarter_driver_qualcomm"]
-asyncio_mode = "auto"
-
-[build-system]
-requires = ["hatchling", "hatch-vcs", "hatch-pin-jumpstarter"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.hooks.pin_jumpstarter]
-name = "pin_jumpstarter"
-
-[dependency-groups]
-dev = [
-    "pytest-cov>=6.0.0",
-    "pytest>=8.3.3",
-    "pytest-asyncio>=0.0.0",
-]
