+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/__init__.py
@@ -1,0 +1,1 @@
+"""QDL platform firmware flashing for Qualcomm automotive SoCs."""
