+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/driver_test.py
@@ -3,10 +3,10 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from jumpstarter_driver_qualcomm.driver import QualcommFlasher
-from jumpstarter_driver_qualcomm.executor import execute_manifest
-from jumpstarter_driver_qualcomm.schema import SleepStep, load_firmware_manifest
-from jumpstarter_driver_qualcomm.soc_profiles import SA8775P
+from jumpstarter_driver_ridesx.qdl.driver import QualcommFlasher
+from jumpstarter_driver_ridesx.qdl.executor import execute_manifest
+from jumpstarter_driver_ridesx.qdl.schema import SleepStep, load_firmware_manifest
+from jumpstarter_driver_ridesx.qdl.soc_profiles import SA8775P
 
 from jumpstarter.common.exceptions import ConfigurationError
