+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/examples/manifests/cs4.yaml
