+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/examples/manifests/cs5.yaml
