+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/examples/manifests/es22.yaml
