+++ python/packages/jumpstarter-driver-ridesx/jumpstarter_driver_ridesx/qdl/firmware_id.py
